backend/requirements.txt
# --- Core API Framework ---
fastapi>=0.110.0
uvicorn[standard]>=0.28.0
pydantic-settings>=2.2.0
email-validator>=2.1.1
python-multipart>=0.0.9

# --- Database & ORM ---
sqlalchemy>=2.0.29
alembic>=1.13.1
psycopg2-binary>=2.9.9

# --- Security & Auth (Self-Hosted JWT) ---
python-jose[cryptography]>=3.3.0
passlib==1.7.4
bcrypt==3.2.2
